+++ Sources/ConsolePerseusLogger/TheOne.swift
@@ -59,9 +59,10 @@ public typealias log = PerseusLogger // In SPM package should be not public exce
 public typealias ConsoleObject = (subsystem: String, category: String)
 public typealias LocalTime = (date: String, time: String, timeUTC: TimeInterval)
 public typealias PIDandTID = (pid: String, tid: String) // PID and Thread ID.
+public typealias Directives = (fileName: String, line: UInt)
 
 public typealias MessageDelegate = (
-    (String, PerseusLogger.Level, LocalTime, PIDandTID, PerseusLogger.User) -> Void
+    (String, PerseusLogger.Level, LocalTime, PIDandTID, PerseusLogger.User, Directives) -> Void
 )
 
 public class PerseusLogger {
@@ -81,7 +82,7 @@ public class PerseusLogger {
     public enum Output: String, Decodable, CaseIterable {
         case standard // In Use: Swift.print("").
         case consoleapp // In Use: Logger structure from iOS 14.0, macOS 11.0, NSLog otherwise.
-        case custom // In Use: customActionOnMessage?(_:_:_:_:_:).
+        case custom // In Use: customActionOnMessage?(_:_:_:_:_:_:).
     }
 
     // log.message("Notification...", .notice, .custom, .enduser)
@@ -263,9 +264,9 @@ public class PerseusLogger {
         // Path.
 
         let withDirectives = (format == .full) ? true : directives && (format != .textonly)
+        let fileName = (file.description as NSString).lastPathComponent
 
         if withDirectives {
-            let fileName = (file.description as NSString).lastPathComponent
             message = "\(text()), file: \(fileName), line: \(line)"
         } else {
             message = "\(text())"
@@ -297,7 +298,8 @@ public class PerseusLogger {
         // Print.
 
         if oput == .custom {
-            customActionOnMessage?(message, type, localTime, idtuple, user)
+            let directives: Directives = (fileName: fileName, line: line)
+            customActionOnMessage?(message, type, localTime, idtuple, user, directives)
         } else {
             print(message, type, oput)
         }